History branch

+++ src/main/java/com/example/test/DTO/HistoryDTO.java
@@ -9,7 +9,7 @@ public class HistoryDTO {
 
     private Long id;
     private Date transactionDate;
-    private Double receivedAmount;
+    private Long receivedAmount;
     private Double totalAmount;
     private String transactionType;
     private String description;
@@ -48,11 +48,11 @@ public class HistoryDTO {
         this.transactionDate = transactionDate;
     }
 
-    public Double getReceivedAmount() {
+    public Long getReceivedAmount() {
         return receivedAmount;
     }
 
-    public void setReceivedAmount(Double receivedAmount) {
+    public void setReceivedAmount(Long receivedAmount) {
         this.receivedAmount = receivedAmount;
     }
